解决spring源码在windowns系统无法构建成功的问题

+++ spring-r2dbc/src/main/java/org/springframework/r2dbc/core/UpdatedRowsFetchSpec.java
@@ -30,6 +30,6 @@ public interface UpdatedRowsFetchSpec {
 	 * Get the number of updated rows.
 	 * @return a Mono emitting the number of updated rows
 	 */
-	Mono<Long> rowsUpdated();
+	Mono<Integer> rowsUpdated();
 
 }
